pyproject.toml
[project]
name = "gcpu-albert-hackathon"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "mintyfrankie", email = "77310871+mintyfrankie@users.noreply.github.com" },
]
dependencies = [
    "streamlit>=1.40.0",
    "langchain-core>=0.3.15",
    "python-dotenv>=1.0.1",
    "langchain-google-genai>=2.0.4",
    "langchain>=0.3.7",
    "langgraph>=0.2.45",
    "pydantic>=2.9.2",
    "pytelegrambotapi>=4.23.0",
    "pydantic-settings>=2.6.1",
    "streamlit-js-eval>=0.1.7",
    "pillow>=11.0.0",
    "google-cloud-vision>=3.8.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["ipykernel>=6.29.5", "pytest>=8.3.3", "watchdog>=6.0.0"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/backend", "src/telegram_worker", "src/web"]

[tool.rye.scripts]
telegram = "python src/telegram_worker/app.py"
web = "streamlit run src/web/app.py"
